requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
validators==0.22.0

